Add configuration option with flags.
If no flags are used, the default config file will be read (revolver.yml).
If the build(b) or run(r) flags are not used, a config file will be read
based on the configFile(c) flag or its default value. Otherwise the config
will be constructed with the specified flags, and the config file will be
ignored. It is possible to add multiple excludeDir(ed), pattern(p), exclude(e)
and build(b) flags (ex: `revolver -b "echo 1" -b "echo 2"'`).